.port most of WebCore's current 'Length' API to KHTML - the only conversion  I did not do is width()/minWidth() => calcValue()/calcMinValue()  as it is a bit tricky. .merge some work bringing better precision to percentage calculations.
Combined with Web Fonts, this let us get rid of the red square+purple 'X'
at the top right of acid3.

CCBUG: 156947

svn path=/trunk/KDE/kdelibs/; revision=965865

+++ khtml/xml/dom_stringimpl.cpp
@@ -305,7 +305,7 @@ static Length parseLength(const QChar *s, unsigned int l)
             const QChar* next = s+i;
 
             if (*next == '%')
-                return Length(r, Percent);
+                return Length(static_cast<double>(r), Percent);
 
             if (*next == '*') 
                 return Length(r, Relative);
